class-43/demo/snackapi/pyproject.toml
[tool.poetry]
name = "snackapi"
version = "0.1.0"
description = ""
authors = ["Ahmad Alawad <ahmad.alawad.sf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
